+++ src/angular-parse/angular.ts
@@ -1,0 +1,2 @@
+export * from './angular/compiler';
+export * from './angular/facade';